check if symlink doesn't exist before adding

+++ add_sh_symlinks.sh
@@ -1,6 +1,15 @@
 #!/usr/bin/env zsh
 
 script_path=${0:a:h}
-ln -sf $script_path/zsh_aliases $HOME/.zsh_aliases
-ln -sf $script_path/zshrc $HOME/.zshrc
-ln -sf $script_path/zsh_aliases $HOME/.bash_aliases
+if [ ! -f $HOME/.zsh_aliases ]; then
+    ln -sf $script_path/zsh_aliases $HOME/.zsh_aliases
+fi
+if [ ! -f $HOME/.zshrc ]; then
+    ln -sf $script_path/zshrc $HOME/.zshrc
+fi
+if [ ! -f $HOME/.bash_aliases ]; then
+    ln -sf $script_path/zsh_aliases $HOME/.bash_aliases
+fi
+if [ ! -f $HOME/.bashrc ]; then
+    ln -sf $script_path/bashrc $HOME/.bashrc
+fi
